Criação do RoomStatus para alterar o status dos quartos

+++ src/main/java/com/eclipsehotel/desafio/services/RoomService.java
@@ -1,6 +1,7 @@
 package com.eclipsehotel.desafio.services;
 
 import com.eclipsehotel.desafio.models.Room;
+import com.eclipsehotel.desafio.models.RoomStatus;
 import com.eclipsehotel.desafio.repositorys.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -35,6 +36,21 @@ public class RoomService {
     }
 
     public void deleteRoom(Long id) {
-        roomRepository.deleteById(id);
+        if (roomRepository.existsById(id)) {
+            roomRepository.deleteById(id);
+        }
+    }
+
+    public List<Room> getOccupiedRooms() {
+        return roomRepository.findByStatus(RoomStatus.OCCUPIED);
+    }
+
+    public Room updateRoomStatus(Long id, RoomStatus status) {
+        Room room = roomRepository.findById(id).orElse(null);
+        if (room != null) {
+            room.setStatus(status);
+            return roomRepository.save(room);
+        }
+        return null; // ou lançar uma exceção se preferir
     }
 }
